# docs/swagger.yaml
definitions:
  controller.CommonError:
    properties:
      error:
        type: string
    type: object
  controller.CommonResponse:
    properties:
      message:
        type: string
    type: object
  controller.CreateEmployeesRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  controller.SignInRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  controller.SignUpRequest:
    properties:
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  controller.TokenPairResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
info:
  contact:
    email: test@test.com
    name: MyProject API Service
    url: http://test.com
  title: MyProject API
paths:
  /api/employees:
    get:
      description: Получения списка всех пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.CommonResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: Получение данных пользователя
      tags:
      - Employees
  /api/employees/{id}:
    delete:
      description: Удаление данных пользователя по ID
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Удаление
      tags:
      - Employees
    get:
      description: Получения информации о пользователе по ID
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: вставьте refresh token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: Получение данных пользователя по ID
      tags:
      - Employees
    put:
      description: Обновление данных пользователя по ID
      parameters:
      - description: информация о пользователе
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.CreateEmployeesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Обновление данных пользователя
      tags:
      - Employees
  /auth/refresh:
    get:
      description: Обновить пару токенов
      parameters:
      - description: вставьте refresh token
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Обновить пару токенов
      tags:
      - Auth
  /auth/sign-in:
    post:
      description: Вход в аккаунт
      parameters:
      - description: Данные для входа пользователя
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Вход
      tags:
      - Auth
  /auth/sign-up:
    post:
      description: Создание нового аккаунта
      parameters:
      - description: Информация о новом аккаунте
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Регистрация
      tags:
      - Auth
  /ping:
    get:
      description: Проверка сервиса
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
      summary: Health-check
      tags:
      - Ping
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"